+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.0.2] - 2023-04-15
+
+### Added
+
+- Implemented Amount.Int64 method.
+- Implemented Amount.Float64 method.
+
+### Changed
+
+- Reviewed and improved documentation.
+
 ## [0.0.1] - 2023-04-13
 
 ### Added